+++ Shared/Models/Cell.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Collections.Generic;
+
+namespace FreeTabletop.Shared.Models
+{
+    public class Cell
+    {
+        public int[] Position { get; set; } = {0, 0};
+
+        public bool IsBlackout = true;
+    }
+}