+++ changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - tabletop loading animation
 - tabletop image loading error notification
 - image loading sfx
+- initial fog of war functionality
 
 ### Fixed